Move era detection out of the transport base and name the 404 expiry rule
The transport base grew past its length budget with the lower branches'
fifth-round changes, and send_http_request past its branching budget with
the well-formed -32601 distinction. The probe-failure verdicts move into
HttpTransportBase::EraDetection unchanged, and the 404 session-expiry
decision becomes one predicate shared by the response and raise_error paths.
